update eventvolunteerlist component and use in event page
