+++ src/components/Event/index.tsx
@@ -1,10 +1,10 @@
 'use client';
 import { EventResponse } from '@/types/dataModel/event';
 import { EventVolunteerResponse } from '@/types/dataModel/eventVolunteer';
-import { Box, ListItemButton, ListItemText, Typography } from '@mui/material';
+import { Box, Typography } from '@mui/material';
 import React from 'react';
 import RoleIconList from '../RoleIconList';
-import { useRouter } from 'next/navigation';
+import EventVolunteerList from '../EventVolunteerList';
 
 interface EventProps {
   event: EventResponse;
@@ -15,7 +15,6 @@ export default function Event({
   event,
   eventVolunteers,
 }: EventProps): React.ReactElement {
-  const router = useRouter();
   // Format the date to a localized string
   const formatDate = (date: Date): string => {
     const option: Intl.DateTimeFormatOptions = {
@@ -64,28 +63,7 @@ export default function Event({
       </Typography>
       {/* TODO remove the below and replace with EventVolounteerList component */}
 
-      {eventVolunteers.length ? (
-        eventVolunteers.map((ev) => (
-          <ListItemButton
-            key={ev._id}
-            onClick={() => router.push(`/volunteers/${ev.volunteer._id}`)}
-            sx={{ pl: 0 }}
-          >
-            <ListItemText
-              primary={`${ev.volunteer.firstName} ${ev.volunteer.lastName}`}
-              secondary={ev.volunteer.email}
-              primaryTypographyProps={{ variant: 'h5' }}
-            />
-            <Box>
-              <RoleIconList roles={[ev.role]}></RoleIconList>
-            </Box>
-          </ListItemButton>
-        ))
-      ) : (
-        <Typography variant="h5" pt={2}>
-          No volunteers yet!
-        </Typography>
-      )}
+      <EventVolunteerList eventVolunteers={eventVolunteers} />
     </Box>
   );
 }
